components/image_embedding/requirements.txt
git+https://github.com/ml6team/fondant@main
gcsfs==2023.4.0
Pillow==9.4.0
torch==2.0.0
transformers==4.28.0